SettingsDaemon: lookup schemas recursive
The docs on  SettingsSchemaSource get_default say
```
The returned source may actually consist of multiple schema sources from different directories, depending on which directories were given in `XDG_DATA_DIRS` and `GSETTINGS_SCHEMA_DIR`. For this reason, all lookups performed against the default source should probably be done recursively.
```

Without this in NixOS it will not work as intended, as we
have multiple entries in XDG_DATA_DIRS.
(we also happen to use an ubuntu patched gsd for pantheon)