vapi: Convert `elif` blocks to `if` blocks
This is a trivial change that (hopefully) shouldn't change any behavior.
We want to clean up the code base with script but
the script does not take care of `elif` blocks yet,
luckily we don't have many `elif` blocks so we simply
take care of them before we run the script.